backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
requests>=2.31.0
python-multipart>=0.0.9

# Modulo Identificazione — NER ibrido
presidio-analyzer>=2.2.0
spacy>=3.7.0
lingua-language-detector>=2.0.0
# After installing, download the spaCy models for the languages you need:
#   python -m spacy download it_core_news_sm   (Italian  — recommended)
#   python -m spacy download en_core_web_sm    (English)
#   python -m spacy download fr_core_news_sm   (French)
#   python -m spacy download de_core_news_sm   (German)
